+++ Assets/Plugins/iOS/UnityARKit/UnityPointCloudExample.cs
@@ -1,9 +1,7 @@
 ﻿using System;
 using UnityEngine;
 using UnityEngine.XR.iOS;
-using System.Collections;
 using System.Collections.Generic;
-using System.Runtime.InteropServices;
 
 public class UnityPointCloudExample : MonoBehaviour
 {
@@ -11,10 +9,11 @@ public class UnityPointCloudExample : MonoBehaviour
     public GameObject PointCloudPrefab = null;
     private List<GameObject> pointCloudObjects;
     private Vector3[] m_PointCloudData;
+    private bool isInitialized;
+    private bool eventsSubscribed;
 
     public void Start()
     {
-        UnityARSessionNativeInterface.ARFrameUpdatedEvent += ARFrameUpdated;
         if (PointCloudPrefab != null)
         {
             pointCloudObjects = new List<GameObject> ();
@@ -23,6 +22,61 @@ public class UnityPointCloudExample : MonoBehaviour
                 pointCloudObjects.Add (Instantiate (PointCloudPrefab));
             }
         }
+        isInitialized = true;
+        SubscribeToEvents ();
+    }
+
+    public void OnEnable()
+    {
+        if (isInitialized)
+        {
+            SubscribeToEvents ();
+        }
+    }
+
+    public void OnDisable()
+    {
+        UnsubscribeFromEvents ();
+    }
+
+    public void OnDestroy()
+    {
+        UnsubscribeFromEvents ();
+        if (pointCloudObjects != null)
+        {
+            foreach (GameObject pointCloudObject in pointCloudObjects)
+            {
+                if (pointCloudObject != null)
+                {
+                    Destroy (pointCloudObject);
+                }
+            }
+            pointCloudObjects.Clear ();
+            pointCloudObjects = null;
+        }
+        m_PointCloudData = null;
+    }
+
+    private void SubscribeToEvents()
+    {
+        if (eventsSubscribed)
+        {
+            return;
+        }
+
+        UnityARSessionNativeInterface.ARFrameUpdatedEvent += ARFrameUpdated;
+        eventsSubscribed = true;
+    }
+
+    private void UnsubscribeFromEvents()
+    {
+        if (!eventsSubscribed)
+        {
+            return;
+        }
+
+        UnityARSessionNativeInterface.ARFrameUpdatedEvent -= ARFrameUpdated;
+        eventsSubscribed = false;
     }
 
     public void ARFrameUpdated(UnityARCamera camera)